+++ x/gamm/pool-models/stableswap/msgs_test.go
@@ -34,6 +34,7 @@ func TestMsgCreateStableswapPool(t *testing.T) {
 			Sender:               addr1,
 			PoolParams:           poolParams,
 			InitialPoolLiquidity: testPoolAsset,
+			ScalingFactors:       []uint64{1, 1},
 			FuturePoolGovernor:   "",
 		}
 
@@ -109,6 +110,14 @@ func TestMsgCreateStableswapPool(t *testing.T) {
 					SwapFee: sdk.NewDecWithPrec(-1, 2),
 					ExitFee: sdk.NewDecWithPrec(0, 0),
 				}
+				return msg
+			}),
+			expectPass: false,
+		},
+		{
+			name: "scaling factors with invalid lenght",
+			msg: createMsg(func(msg stableswap.MsgCreateStableswapPool) stableswap.MsgCreateStableswapPool {
+				msg.ScalingFactors = []uint64{1, 2, 3}
 				return msg
 			}),
 			expectPass: false,